Push data to the front of the list instead
Instead of appending to the end of the list `O(n)` we should push to
the front which is `O(1)`. This requires us to `Enum.reverse/1` in
`Sentry.Context.get_all/1` (which is `O(n)`) but this method should
be called as little as possible, so the gain should be visible in
breadcrumb heavy applications.

Thanks to that change we also got simple
`Sentry.Context.pop_breadcrumb/0` method to remove head breadcrumb.