allow users to configure maximum number of breadcrumbs
This configuration option allows a user to specify the maximum number of
breadcrumbs when posting events to Sentry. The default value is 100.

This is helpful in situations in which long running processes can add a
large number of breadcrumbs resulting in failures to create Sentry
events when the message becomes too large.

Fixes #407