Improve DSN parsing and Endpoint building
We can now parse DSN's that have more in the path as just the ID.
Also we build the Endpoint again with that path, followed by api/<id>/store.
This change brings the DSN parsing and Endpoint building up to par with
how the Ruby SDK is doing it.

Close #509